requirements.txt
PyMuPDF~=1.25.3
pillow~=11.1.0